example/requirements.txt
requests = "^2.24.0"
oauthlib = "^3.1.0"
requests-oauthlib = "^1.3.0"
python-dotenv = "^0.14.0"
